pyproject.toml
[project]
name = "omnic"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.2.2",
    "python-telegram-bot>=22.6",
]
